requirements.txt
dj-database-url==0.5.0
Django==2.1.2
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.9.0
gunicorn==19.9.0
Markdown==3.0.1
psycopg2==2.7.5
pytz==2018.6
whitenoise==4.1
dj-database-url==0.5.0
Django==2.1.2
django-cors-headers==3.0.1
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.9.0
gunicorn==19.9.0
Markdown==3.0.1
psycopg2==2.7.5
pytz==2018.6
whitenoise==4.1
dj-database-url==0.5.0
Django==2.1.2
django-cors-headers==3.0.1
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Markdown==3.0.1
psycopg2==2.7.5
psycopg2-binary==2.8.2
PyJWT==1.7.1
pytz==2018.6
sqlparse==0.3.0
whitenoise==4.1
dj-database-url==0.5.0
Django==2.1.2
django-cors-headers==3.0.1
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Markdown==3.0.1
psycopg2==2.7.5
psycopg2-binary==2.8.2
PyJWT==1.7.1
pytz==2018.6
sqlparse==0.3.0
whitenoise==4.1
dj-database-url==0.5.0
Django==2.1.2
django-cors-headers==3.0.1
django-filter==2.0.0
django-heroku==0.3.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Markdown==3.0.1
Pillow==6.0.0
psycopg2==2.7.5
psycopg2-binary==2.8.2
PyJWT==1.7.1
pytz==2018.6
sqlparse==0.3.0
whitenoise==4.1
